Release and Debug profiles separated
Make use of Release and Debug separate profiles.

Since testing introduced, it is quite necessary to disable all the
optimizations to be able to debug test cases.

But it is still interesting to see an optimized version of assembler
listings while not debugging.

So there, optimized and not optimized profiles were separated.